+++ packages/1xdev-ui/__test__/heroMessage.spec.tsx
@@ -1,6 +1,6 @@
 import { render, screen, act } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { HeroMessage } from '../../../../src';
+import { AnimatedHeading } from '../src';
 
 jest.mock('framer-motion', () => ({
   motion: {
@@ -11,16 +11,22 @@ jest.mock('framer-motion', () => ({
 
 jest.useFakeTimers();
 
-describe('HeroMessage Component', () => {
+describe('AnimatedHeading Component', () => {
+  const animatedHeadingProps = {
+    phrase: "We {keyWords} the future while being {keyDescriptions}",
+    keyWords: ["{engineer}", '{create}', '{code}'],
+    keyDescriptions: ['{humans}', '{developers}', '{artists}']
+  }
+
   it('renders the initial phrase correctly', () => {
-    render(<HeroMessage />);
-    const heroMessage = screen.getByTestId('hero message');
-    expect(heroMessage).toBeInTheDocument();
+    render(<AnimatedHeading phrase={animatedHeadingProps.phrase} keyWords={animatedHeadingProps.keyDescriptions} keyDescriptions={animatedHeadingProps.keyDescriptions} />);
+    const animatedAnimatedHeading = screen.getByTestId('hero message');
+    expect(animatedAnimatedHeading).toBeInTheDocument();
     expect(screen.getByTestId('hero message')).toHaveTextContent('We engineer the future while being humans');
   });
 
   it('updates the dynamic words every 3 seconds', () => {
-    render(<HeroMessage />);
+    render(<AnimatedHeading phrase={animatedHeadingProps.phrase} keyWords={animatedHeadingProps.keyDescriptions} keyDescriptions={animatedHeadingProps.keyDescriptions} />);
 
     expect(screen.getByTestId('hero message')).toHaveTextContent('We engineer the future while being humans');
 
@@ -32,7 +38,7 @@ describe('HeroMessage Component', () => {
   });
 
   it('cycles back to the first phrase after all dynamic words have been shown', () => {
-    render(<HeroMessage />);
+    render(<AnimatedHeading phrase={animatedHeadingProps.phrase} keyWords={animatedHeadingProps.keyDescriptions} keyDescriptions={animatedHeadingProps.keyDescriptions} />);
 
     act(() => {
       jest.advanceTimersByTime(9000);
