+++ packages/1xdev-ui/src/index.tsx
@@ -1,2 +1,2 @@
 export * from './ui'
-export * from './pages'
+export * from './utils'
