--- packages/1xdev-ui/src/pages/home/components/index.tsx
@@ -1,2 +1,0 @@
-export * from './heroMessage'
-export * from './renderDynamicPhrase'
