--- packages/1xdev-ui/src/pages/home/homePage.tsx
@@ -1,10 +1,0 @@
-'use client'
-import {HeroMessage} from './components'
-
-export function HomePage() {
-  return (
-    <>
-      <HeroMessage />
-    </>
-  )
-}
