--- packages/1xdev-ui/src/pages/home/index.tsx
@@ -1,2 +1,0 @@
-export * from './homePage'
-export * from './components'
