--- packages/1xdev-ui/src/pages/index.tsx
@@ -1,1 +1,0 @@
-export * from './home'
