+++ packages/1xdev-ui/src/ui/animatedHeading/animatedHeading.module.scss
