+++ packages/1xdev-ui/src/ui/animatedHeading/index.tsx
@@ -1,0 +1,1 @@
+export * from './animatedHeading'
