+++ packages/1xdev-ui/src/ui/index.tsx
@@ -1,2 +1,3 @@
 export * from './button'
 export * from './iconLink'
+export * from './animatedHeading'
